requirements.txt
Flask-WTF == 0.14.2
WTForms == 2.1